enhancement/refactor search-pages * changed divs to boxes * changed h4 to Typography * refactored FormObserver component * refactored scss to mui where it's possible ( ignoring formik and legacy components )
