--- src/components/pages/search-pages/subject-teacher-search/SearchTeacherPage.module.scss
@@ -1,13 +1,0 @@
-@import "src/styles/partials/typography-mixins";
-@import "src/styles/partials/breakpoints";
-
-
-.subject-title{
-   @include h4--semi-bold;
-}
-
-@media screen and (max-width: $breakpoint-tablet) {
-    .subject-title{
-        @include h6--bold;
-    }
-}
